2025-07-fast-track-to-serverless/requirements.txt
databricks-sdk==0.47.0